requirements.txt
black==22.3.0
click==8.1.2
flake8==4.0.1
Flask==2.1.1
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
marshmallow-sqlalchemy==0.28.0
pre-commit==2.18.1
psycopg2==2.9.3
pytz==2022.1
Werkzeug==2.1.1
